xserver libX11 pixman xwininfo mesa plink libXdmcp fontconfig libXau libXext libXft libXmu libXinerama libfontenc libxcb xkeyboard-config  git update Jun 29 2020
xserver          commit b67052742980fd3ec669100048da71e09aa61358
libxcb           commit ee9dfc9a7658e7fe75d27483bb5ed1ba4d1e2c86
libxcb/xcb-proto commit 97fbba25baf241cbcda882739db423b1d36e440a
xkeyboard-config commit 5ea031eb4998726dc5a00ba1fb2253b21cdbe366
libX11           commit 4385a84c4a79c1b60bf12839899a1544198de975
libXdmcp         commit f1bf902c52884e35d48f4d03915f019337f84717
libXext          commit a07b4bb8290d0c1bba7bcecd5bb6896fbe1b169c
libfontenc       commit 8d30ca25b54934e29206da64e80c044d05b14470
libXinerama      commit 56777107b0c517ba58304d6a247bf65f7a27eb96
libXau           commit 42e152c6f2d1bd839e77c5e97f3a509d890c3237
xkbcomp          commit 8995ac58e15595de9158c2d8df9e02b0c925e8f4
pixman           commit 9b49f4e08751885289333fed652bf5e0f45976b4
mkfontscale      commit 79f686765238ddd0e4753f9d4ee62b5a5296610e
xwininfo         commit 0c49f8f2bd56b1e77721e81030ea948386dcdf4e
libXft           commit c418dc7594f98359ae10815f62ee3efc0a511cf8
libXmu           commit 4f0939d69a17d37be414d0de7ff717e92e7e2b11
libxtrans        commit 3b5df889f58a99980a35a7b4a18eb4e7d2abeac4
fontconfig       commit 37c7c748740bf6f2468d59e67951902710240b34
mesa             commit 622c548967a3ba7f7522057e10d696600b1b002c
putty            commit 2762a2025f094bf8ccb75c0bdd7c655bc953f2f1
pthreads         commit 67186579ee3cf81e54d149dce5f5a612f957d2fe
EGL-Registry     commit f636b23410dd4db5055dffbe499f4754013759d5
xorgproto        commit c62e8203402cafafa5ba0357b6d1c019156c9f36

+++ libX11/src/ErrDes.c
@@ -204,9 +204,9 @@ XGetErrorDatabaseText(
 	result.addr = (XPointer)NULL;
     if (!result.addr) {
 	result.addr = (XPointer) defaultp;
-	result.size = strlen(defaultp) + 1;
+	result.size = (unsigned)strlen(defaultp) + 1;
     }
-    (void) strncpy (buffer, (char *) result.addr, nbytes);
+    (void) strncpy (buffer, (char *) result.addr, (size_t)nbytes);
     if (result.size > nbytes) buffer[nbytes-1] = '\0';
     return 0;
 }
