+++ libX11/src/StName.c
@@ -38,7 +38,7 @@ XStoreName (
 {
     return XChangeProperty(dpy, w, XA_WM_NAME, XA_STRING,
 			   8, PropModeReplace, (_Xconst unsigned char *)name,
-			   name ? strlen(name) : 0);
+			   name ? (int) strlen(name) : 0);
 }
 
 int
@@ -49,5 +49,5 @@ XSetIconName (
 {
     return XChangeProperty(dpy, w, XA_WM_ICON_NAME, XA_STRING, 8,
                            PropModeReplace, (_Xconst unsigned char *)icon_name,
-			   icon_name ? strlen(icon_name) : 0);
+			   icon_name ? (int) strlen(icon_name) : 0);
 }
