+++ libX11/src/xcms/LRGB.c
@@ -1293,7 +1293,7 @@ XcmsLRGB_RGB_ParseString(
 	 * Attempt to parse the value portion.
 	 */
 	spec++;
-	n = strlen(spec);
+	n = (int)strlen(spec);
 	if (n != 3 && n != 6 && n != 9 && n != 12) {
 	    return(XcmsFailure);
 	}
@@ -1337,7 +1337,7 @@ XcmsLRGB_RGB_ParseString(
 	/*
 	 * Check for proper prefix.
 	 */
-	if (strncmp(spec, _XcmsRGB_prefix, n) != 0) {
+	if (strncmp(spec, _XcmsRGB_prefix, (size_t)n) != 0) {
 	    return(XcmsFailure);
 	}
