+++ libX11/src/xcms/Luv.c
@@ -152,7 +152,7 @@ CIELuv_ParseString(
     /*
      * Check for proper prefix.
      */
-    if (strncmp(spec, _XcmsCIELuv_prefix, n) != 0) {
+    if (strncmp(spec, _XcmsCIELuv_prefix, (size_t)n) != 0) {
 	return(XcmsFailure);
     }
