+++ libX11/src/xkb/XKBExtDev.c
@@ -651,7 +651,7 @@ _XkbWriteSetDeviceInfo(char *wire,
     if (changes->changed & XkbXI_ButtonActionsMask) {
         int size = changes->num_btns * SIZEOF(xkbActionWireDesc);
 
-        memcpy(wire, (char *) &devi->btn_acts[changes->first_btn], size);
+        memcpy(wire, (char *) &devi->btn_acts[changes->first_btn], (size_t) size);
         wire += size;
     }
     if (changes->changed & XkbXI_IndicatorsMask) {
