+++ libX11/src/xlibi18n/XDefaultIMIF.c
@@ -452,7 +452,7 @@ _WcLookupString(
     } else {
 	*status = XLookupKeySym;
     }
-    mbstowcs(buffer, mb_buf, length);
+    mbstowcs(buffer, mb_buf, (size_t) length);
     XFree(mb_buf);
     return(length);
 }
