+++ libX11/src/xlibi18n/lcPublic.c
@@ -219,7 +219,7 @@ initialize(
      * _XlMapOSLocaleName will return the same string or a substring
      * of name, so strlen(name) is okay
      */
-    if ((len = strlen(name)) < sizeof sinamebuf)
+    if ((len = (int) strlen(name)) < sizeof sinamebuf)
         siname = sinamebuf;
     else
         siname = Xmalloc (len + 1);
